restructure executor folder
