+++ execution/executor/src/lib.rs
@@ -9,11 +9,10 @@ pub mod fuzzing;
 mod logging;
 pub mod metrics;
 #[cfg(test)]
-mod mock_vm;
-#[cfg(test)]
 mod tests;
 
 pub mod block_executor;
 pub mod chunk_executor;
-pub mod components;
 pub mod db_bootstrapper;
+pub mod types;
+pub mod workflow;
