+++ execution/executor/src/tests/chunk_executor_tests.rs
@@ -8,8 +8,10 @@ use crate::{
     block_executor::BlockExecutor,
     chunk_executor::ChunkExecutor,
     db_bootstrapper::{generate_waypoint, maybe_bootstrap},
-    mock_vm::{encode_mint_transaction, MockVM},
-    tests::{self, create_blocks_and_chunks, create_transaction_chunks},
+    tests::{
+        self, create_blocks_and_chunks, create_transaction_chunks,
+        mock_vm::{encode_mint_transaction, MockVM},
+    },
 };
 use aptos_crypto::HashValue;
 use aptos_db::AptosDB;
