+++ execution/executor/src/types/mod.rs
@@ -4,15 +4,7 @@
 
 #![forbid(unsafe_code)]
 
-pub mod apply_chunk_output;
-pub mod block_tree;
-pub mod chunk_commit_queue;
 pub mod in_memory_state_calculator_v2;
 
-pub mod chunk_result_verifier;
-pub mod do_get_execution_output;
-pub mod do_ledger_update;
-pub mod do_state_checkpoint;
 pub mod executed_chunk;
 pub mod partial_state_compute_result;
-pub mod transaction_chunk;
