+++ execution/executor/src/workflow/mod.rs
@@ -4,28 +4,26 @@
 
 #![forbid(unsafe_code)]
 
-use crate::components::{
-    do_ledger_update::DoLedgerUpdate, do_state_checkpoint::DoStateCheckpoint,
-    partial_state_compute_result::PartialStateComputeResult,
-};
+use crate::types::partial_state_compute_result::PartialStateComputeResult;
 use anyhow::Result;
 use aptos_crypto::HashValue;
 use aptos_executor_types::execution_output::ExecutionOutput;
 use aptos_storage_interface::ExecutedTrees;
-use aptos_types::transaction::Transaction;
+use do_ledger_update::DoLedgerUpdate;
+use do_state_checkpoint::DoStateCheckpoint;
 
-pub struct ApplyChunkOutput;
+pub mod do_get_execution_output;
+pub mod do_ledger_update;
+pub mod do_state_checkpoint;
 
-impl ApplyChunkOutput {
-    pub fn apply_chunk(
+pub struct ApplyExecutionOutput;
+
+impl ApplyExecutionOutput {
+    pub fn run(
         chunk_output: ExecutionOutput,
         base_view: &ExecutedTrees,
         known_state_checkpoint_hashes: Option<Vec<Option<HashValue>>>,
-    ) -> Result<(
-        PartialStateComputeResult,
-        Vec<Transaction>,
-        Vec<Transaction>,
-    )> {
+    ) -> Result<PartialStateComputeResult> {
         let (result_state, next_epoch_state, state_checkpoint_output) = DoStateCheckpoint::run(
             chunk_output,
             base_view.state(),
@@ -33,7 +31,7 @@ impl ApplyChunkOutput {
             known_state_checkpoint_hashes,
             /*is_block=*/ false,
         )?;
-        let (ledger_update_output, to_discard, to_retry) =
+        let (ledger_update_output, _to_discard, _to_retry) =
             DoLedgerUpdate::run(state_checkpoint_output, base_view.txn_accumulator().clone())?;
         let output = PartialStateComputeResult::new(
             base_view.state().clone(),
@@ -42,6 +40,6 @@ impl ApplyChunkOutput {
         );
         output.set_ledger_update_output(ledger_update_output);
 
-        Ok((output, to_discard, to_retry))
+        Ok(output)
     }
 }
